Feat : DTO Fix
Former-commit-id: 78b4e45089ba9bed0308d0d9a57e2fbe29e62629

+++ backend/src/main/java/com/a503/churros/dto/news/ArticleInputDTO.java
@@ -9,5 +9,5 @@ import lombok.*;
 @AllArgsConstructor
 @ToString
 public class ArticleInputDTO {
-    private Long articleId;
+    private Integer articleId;
 }
